Pagination done

+++ app/src/main/java/com/example/androidproject/api/model/Restaurants.kt
@@ -1,7 +1,7 @@
 package com.example.androidproject.api.model
 
 data class Restaurants(
-    val total_entries: Long,
+    val total_entries: Int,
     val page: Int,
     val per_page: Int,
     /*val current_page: Int,*/
